fix(ci): unblock ubuntu and windows test jobs
Two independent CI failures, neither in product code:

- The workspace stable-id suite now lives inside the existing
  orchestrator_workspace_roundtrip test binary instead of adding a 43rd
  integration-test target. Each tests/*.rs file links its own full
  editor executable; the extra ~1.5GB debug binary was tipping the
  ubuntu runner into "No space left on device" during cargo test
  --no-run (ld died with SIGBUS).

- test_search_replace_file_glob asserted literal forward-slash paths
  like "src/main.rs (1/1)" against the rendered results tree, but the
  tree shows workspace-relative paths with the platform separator, so
  on Windows the wait_until predicates could never match and the test
  hit its 180s timeout deterministically. Expected paths are now built
  with MAIN_SEPARATOR (same idiom as copy_buffer_path.rs); the typed
  glob stays "src/*.rs" since the glob engine accepts both separators.
