+++ crates/fresh-editor/tests/orchestrator_workspace_roundtrip.rs
@@ -6,6 +6,8 @@
 //! "faithful per-window restore" half of the spec and de-risks the
 //! Stage 1 refactor (moving capture/restore onto `Window`): the
 //! observable round-trip behavior must not change.
+//!
+//! Also hosts the [`stable_id`] suite for durable workspace identity.
 
 use fresh::config::Config;
 use fresh::config_io::DirectoryContext;
@@ -102,5 +104,253 @@ proptest! {
             expected,
             "restored window must reopen exactly the files that were saved"
         );
+    }
+}
+
+/// Durable workspace identity (`stable_id`).
+///
+/// A workspace's on-disk snapshot is keyed by an identity minted once at
+/// window creation — `workspaces/<encoded-root>.<stable_id>.json` — with
+/// the encoded root serving only as a filename-level locator. These tests
+/// pin the identity lifecycle: it survives save → restore cycles without
+/// spawning sibling files, legacy root-keyed files (no id) are adopted and
+/// re-keyed on the next save, duplicate files claiming one directory
+/// resolve to the freshest snapshot, and delete removes every variant.
+///
+/// (Lives in this binary rather than its own `tests/*.rs` target: each
+/// integration-test binary links a full editor executable, which is real
+/// CI disk and link time.)
+mod stable_id {
+    use fresh::config::Config;
+    use fresh::config_io::DirectoryContext;
+    use fresh::model::filesystem::StdFileSystem;
+    use fresh::workspace::{
+        encode_path_for_filename, find_workspace_file_by_root, get_workspaces_dir, Workspace,
+    };
+    use std::path::{Path, PathBuf};
+    use std::sync::Arc;
+
+    fn editor_in(project: &Path, dir_context: &DirectoryContext) -> fresh::app::Editor {
+        let filesystem: Arc<dyn fresh::model::filesystem::FileSystem + Send + Sync> =
+            Arc::new(StdFileSystem);
+        let config = Config {
+            check_for_updates: false,
+            ..Config::default()
+        };
+        fresh::app::Editor::for_test(
+            config,
+            80,
+            24,
+            Some(project.to_path_buf()),
+            dir_context.clone(),
+            fresh::view::color_support::ColorCapability::TrueColor,
+            filesystem,
+            None,
+            None,
+            false,
+            false,
+        )
+        .unwrap()
+    }
+
+    /// All workspace files in the (process-global) workspaces dir whose name
+    /// starts with `project`'s encoded root. Unique temp roots per test keep
+    /// parallel tests from seeing each other's files.
+    fn files_for_root(project: &Path) -> Vec<PathBuf> {
+        let canonical = project.canonicalize().unwrap();
+        let prefix = encode_path_for_filename(&canonical);
+        let dir = get_workspaces_dir().unwrap();
+        match std::fs::read_dir(dir) {
+            Ok(entries) => entries
+                .flatten()
+                .map(|e| e.path())
+                .filter(|p| {
+                    p.file_name()
+                        .and_then(|n| n.to_str())
+                        .is_some_and(|n| n.starts_with(&*prefix) && n.ends_with(".json"))
+                })
+                .collect(),
+            Err(_) => Vec::new(),
+        }
+    }
+
+    fn read_stable_id(path: &Path) -> Option<String> {
+        let val: serde_json::Value = serde_json::from_str(&std::fs::read_to_string(path).ok()?).ok()?;
+        val.get("stable_id")
+            .and_then(|v| v.as_str())
+            .map(str::to_owned)
+    }
+
+    #[test]
+    fn stable_id_survives_restart_without_file_proliferation() {
+        let sandbox = tempfile::tempdir().unwrap();
+        let project = sandbox.path().join("project");
+        std::fs::create_dir_all(&project).unwrap();
+        let project = project.canonicalize().unwrap();
+        std::fs::write(project.join("a.txt"), "x").unwrap();
+        let dir_context = DirectoryContext::for_testing(&sandbox.path().join("data-home"));
+
+        // Editor 1: open a file, save twice (a checkpoint plus a quit save).
+        let first_id;
+        {
+            let mut e1 = editor_in(&project, &dir_context);
+            e1.open_file(&project.join("a.txt")).unwrap();
+            e1.save_workspace().unwrap();
+            e1.save_workspace().unwrap();
+            let files = files_for_root(&project);
+            assert_eq!(files.len(), 1, "repeated saves reuse one file: {files:?}");
+            first_id = read_stable_id(&files[0]).expect("saved workspace carries a stable_id");
+        }
+
+        // Editor 2 ("restart"): restore, then save again — the identity must
+        // continue, not fork a sibling file with a freshly minted id.
+        let mut e2 = editor_in(&project, &dir_context);
+        assert!(e2.try_restore_workspace().unwrap());
+        assert_eq!(
+            e2.active_window().stable_id,
+            first_id,
+            "the restored window adopts the persisted identity"
+        );
+        e2.save_workspace().unwrap();
+
+        let files = files_for_root(&project);
+        assert_eq!(
+            files.len(),
+            1,
+            "a restart must not mint a sibling workspace file: {files:?}"
+        );
+        assert_eq!(
+            read_stable_id(&files[0]).as_deref(),
+            Some(first_id.as_str())
+        );
+    }
+
+    #[test]
+    fn legacy_root_keyed_file_is_adopted_and_rekeyed_on_save() {
+        let sandbox = tempfile::tempdir().unwrap();
+        let project = sandbox.path().join("project");
+        std::fs::create_dir_all(&project).unwrap();
+        let project = project.canonicalize().unwrap();
+        std::fs::write(project.join("legacy.txt"), "x").unwrap();
+        let dir_context = DirectoryContext::for_testing(&sandbox.path().join("data-home"));
+
+        // A pre-stable-id workspace file at the legacy root-keyed name,
+        // produced by capturing a real layout and stripping the id.
+        let legacy_path = {
+            let mut e = editor_in(&project, &dir_context);
+            e.open_file(&project.join("legacy.txt")).unwrap();
+            let mut ws = e.capture_workspace();
+            ws.stable_id = None;
+            ws.label = Some("legacy-label".to_string());
+            ws.save().unwrap();
+            let files = files_for_root(&project);
+            assert_eq!(files.len(), 1);
+            assert_eq!(
+                files[0],
+                fresh::workspace::get_workspace_path(&project).unwrap(),
+                "an id-less snapshot lands at the legacy root-keyed name"
+            );
+            files[0].clone()
+        };
+
+        // A fresh editor finds the legacy file by root...
+        let mut e = editor_in(&project, &dir_context);
+        assert!(e.try_restore_workspace().unwrap());
+        let adopted = e.active_window().stable_id.clone();
+        assert!(!adopted.is_empty());
+
+        // ...and the next save re-keys it: the id-keyed file appears, the
+        // legacy file is retired, and lookup keeps resolving.
+        e.save_workspace().unwrap();
+        assert!(
+            !legacy_path.exists(),
+            "the superseded legacy file is retired on save"
+        );
+        let files = files_for_root(&project);
+        assert_eq!(
+            files.len(),
+            1,
+            "exactly the re-keyed file remains: {files:?}"
+        );
+        assert_eq!(read_stable_id(&files[0]).as_deref(), Some(adopted.as_str()));
+
+        let loaded = Workspace::load(&project)
+            .unwrap()
+            .expect("loadable by root");
+        assert_eq!(
+            loaded.stable_id.as_deref(),
+            Some(adopted.as_str()),
+            "the re-keyed snapshot carries the adopted identity"
+        );
+    }
+
+    #[test]
+    fn duplicate_files_for_one_root_resolve_to_freshest_snapshot() {
+        let sandbox = tempfile::tempdir().unwrap();
+        let project = sandbox.path().join("project");
+        std::fs::create_dir_all(&project).unwrap();
+        let project = project.canonicalize().unwrap();
+
+        let dir = get_workspaces_dir().unwrap();
+        std::fs::create_dir_all(&dir).unwrap();
+        let encoded = encode_path_for_filename(&project);
+
+        let mut old = Workspace::new(project.clone());
+        old.label = Some("old".to_string());
+        old.saved_at = 100;
+        std::fs::write(
+            dir.join(format!("{encoded}.json")),
+            serde_json::to_vec(&old).unwrap(),
+        )
+        .unwrap();
+
+        let mut new = Workspace::new(project.clone());
+        new.label = Some("new".to_string());
+        new.saved_at = 200;
+        new.stable_id = Some("ws-test-1".to_string());
+        std::fs::write(
+            dir.join(format!("{encoded}.ws-test-1.json")),
+            serde_json::to_vec(&new).unwrap(),
+        )
+        .unwrap();
+
+        let loaded = Workspace::load(&project).unwrap().expect("resolvable");
+        assert_eq!(
+            loaded.label.as_deref(),
+            Some("new"),
+            "arbitration picks the freshest snapshot, not the stale legacy file"
+        );
+
+        // Deleting the workspace removes every variant, so a killed workspace
+        // can't resurrect from the stale duplicate.
+        Workspace::delete(&project).unwrap();
+        assert!(files_for_root(&project).is_empty());
+        assert!(find_workspace_file_by_root(&project).unwrap().is_none());
+    }
+
+    /// An encoded root that is a *prefix* of another's (`/a` vs `/a.b`) must
+    /// not cross-match: lookup verifies the `working_dir` recorded inside each
+    /// candidate, so `a.b`'s files never count for `a`.
+    #[test]
+    fn prefix_colliding_roots_do_not_cross_match() {
+        let sandbox = tempfile::tempdir().unwrap();
+        let short = sandbox.path().join("proj");
+        let long = sandbox.path().join("proj.aux");
+        std::fs::create_dir_all(&short).unwrap();
+        std::fs::create_dir_all(&long).unwrap();
+        let short = short.canonicalize().unwrap();
+        let long = long.canonicalize().unwrap();
+
+        let mut ws_long = Workspace::new(long.clone());
+        ws_long.label = Some("long-root".to_string());
+        ws_long.stable_id = Some("ws-test-long".to_string());
+        ws_long.save().unwrap();
+
+        assert!(
+            Workspace::load(&short).unwrap().is_none(),
+            "the short root must not resolve to the longer root's file"
+        );
+        let loaded = Workspace::load(&long).unwrap().expect("long root loads");
+        assert_eq!(loaded.label.as_deref(), Some("long-root"));
     }
 }
